create text/test_text_methods.py
